ridehistory update

+++ app/src/main/java/com/example/cabme/riders/RideRequest.java
@@ -7,12 +7,16 @@ import androidx.annotation.NonNull;
 import com.example.cabme.User;
 import com.example.cabme.maps.CostAlgorithm;
 import com.example.cabme.maps.JsonParser;
+import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
+import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.firestore.CollectionReference;
 import com.google.firebase.firestore.DocumentReference;
+import com.google.firebase.firestore.DocumentSnapshot;
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.GeoPoint;
 import com.google.firebase.firestore.Query;
@@ -51,6 +55,7 @@ public class RideRequest {
     private transient FirebaseFirestore firebaseFirestore;
     private transient CollectionReference collectionReference;
     private transient DocumentReference documentReference;
+    private transient DatabaseReference databaseReference;
     private String firebaseCollectionName = "testrequests";
 
     /* Keys */
@@ -79,21 +84,24 @@ public class RideRequest {
 
     /**
      * This contructs a ride request with specificed rider ID.
+     *
      * @param reqUserID the user ID which is also the request ID of the ride request
      */
-    public RideRequest(String reqUserID){
+    public RideRequest(String reqUserID) {
         UIDrider = reqUserID;
+        initializeFireBase();
     }
 
     /**
      * This constructs a ride request with the specified start and end location and rider ID.
+     *
      * @param startGeo start location of the ride request
-     * @param endGeo end location of the ride request
+     * @param endGeo   end location of the ride request
      * @param UIDrider ID of the rider
-     * @param API_KEY the Google API key
+     * @param API_KEY  the Google API key
      */
     public RideRequest(GeoPoint startGeo, GeoPoint endGeo,
-                       String UIDrider, String API_KEY, Double rideCost){
+                       String UIDrider, String API_KEY, Double rideCost) {
         setGiven(startGeo, endGeo, UIDrider, API_KEY);
         setParsedGeoPoints();
         setRideCost(rideCost);
@@ -104,13 +112,14 @@ public class RideRequest {
     /**
      * This method sets the given variables.
      * This method is called in the RideRequest() method.
+     *
      * @param startGeo start location of the ride request
-     * @param endGeo end location of the ride request
+     * @param endGeo   end location of the ride request
      * @param UIDrider ID of the rider
-     * @param API_KEY Google API key
+     * @param API_KEY  Google API key
      */
     public void setGiven(GeoPoint startGeo, GeoPoint endGeo,
-                         String UIDrider, String API_KEY){
+                         String UIDrider, String API_KEY) {
         this.API_KEY = API_KEY;
         this.startGeo = startGeo;
         this.endGeo = endGeo;
@@ -120,9 +129,10 @@ public class RideRequest {
     /**
      * This method sets the cost of the ride request.
      * This method is called in the RideRequest method.
+     *
      * @param rideCost the cost of the ride
      */
-    private void setRideCost(Double rideCost){
+    private void setRideCost(Double rideCost) {
         this.rideCost = rideCost;
     }
 
@@ -130,7 +140,7 @@ public class RideRequest {
      * This method uses the JsonParser Class to parse the start and end location and sets the
      * JSON parsed information of the ride request to its appropriate variable.
      */
-    private void setParsedGeoPoints(){
+    private void setParsedGeoPoints() {
         jsonParser = new JsonParser(startGeo, endGeo, API_KEY);
         this.distanceText = jsonParser.getDistanceText();
         this.distanceValue = jsonParser.getDistanceValue();
@@ -138,30 +148,25 @@ public class RideRequest {
         this.durationValue = jsonParser.getDurationValue();
         this.endAddress = jsonParser.getEndAddress();
         this.startAddress = jsonParser.getStartAddress();
-
-        Log.wtf("newrr", "distext: " + distanceText);
-        Log.wtf("newrr", "disvalue: " + distanceValue);
-        Log.wtf("newrr", "durtext: " + durationText);
-        Log.wtf("newrr", "durvalue: " + durationValue);
-        Log.wtf("newrr", "start: " + startAddress);
-        Log.wtf("newrr", "end: " + endAddress);
     }
 
     /**
      * This method initializes the FireBase and and the collection and document reference.
      * This method is called in the constructors.
      */
-    private void initializeFireBase(){
+    private void initializeFireBase() {
         firebaseFirestore = FirebaseFirestore.getInstance();
         collectionReference = firebaseFirestore.collection(firebaseCollectionName);
         documentReference = firebaseFirestore.collection(firebaseCollectionName).document(UIDrider);
+        databaseReference = FirebaseDatabase.getInstance().getReference();
+
     }
 
     /**
      * This method put the ride request information from the variables to a document in the Firebase
      * collection where the drivers can view each riders' ride request.
      */
-    private void putInFirebaseCollection(){
+    private void putInFirebaseCollection() {
         HashMap<String, Object> newRideRequest = new HashMap<>();
         newRideRequest.put("UIDdriver", UIDdriver);
         newRideRequest.put("UIDrider", UIDrider);
@@ -180,18 +185,53 @@ public class RideRequest {
                 .document(UIDrider)
                 .set(newRideRequest)
                 .addOnSuccessListener(aVoid -> Log.d(TAG, "Ride request added "))
-                .addOnFailureListener(e -> Log.d(TAG, "Ride request unable to be added "+ e.toString()));
+                .addOnFailureListener(e -> Log.d(TAG, "Ride request unable to be added " + e.toString()));
     }
 
     /**
      * This method removes a the ride request tied to a user's ID in the Firebase collection where
      * the drivers can view each riders' ride request. .
      */
-    public void removeRequest(){
-        initializeFireBase();
+    public void removeRequest() {
+        String DOCID = FirebaseDatabase.getInstance().getReference("ridehistory").push().getKey();
+        DocumentReference ridehistoryRef = firebaseFirestore
+                .collection("users")
+                .document(UIDrider)
+                .collection("ridehistory")
+                .document(DOCID);
+
         documentReference
-                .delete()
-                .addOnSuccessListener(aVoid -> Log.d(TAG, "Ride request deleted "))
-                .addOnFailureListener(e -> Log.d(TAG, "Ride request unable to be deleted "+ e.toString()));
+                .get()
+                .addOnCompleteListener(task -> {
+                    if (task.isSuccessful()) {
+                        DocumentSnapshot documentSnapshot = task.getResult();
+                        if (documentSnapshot != null) {
+                            ridehistoryRef.set(documentSnapshot.getData())
+                                    .addOnSuccessListener(aVoid -> {
+                                            Log.d(TAG, "DocumentSnapshot successfully written!");
+                                            documentReference
+                                                    .delete()
+                                                    .addOnFailureListener(e -> Log.d(TAG, "Ride request unable to be deleted "+ e.toString()))
+                                                    .addOnSuccessListener(aVoid1 -> Log.d(TAG, "Ride request deleted "));
+                                        }
+                                    )
+                                    .addOnFailureListener(e -> Log.w(TAG, "Error writing document", e));
+                        }
+                        else {
+                            Log.d(TAG, "No such document");
+                        }
+                    }
+                });
+    }
+
+    public void updateRideStatus(String status){
+        // GETS PERMISSION ERROR
+        databaseReference
+                .child(firebaseCollectionName)
+                .child(UIDrider)
+                .child("rideStatus")
+                .setValue(status)
+                .addOnSuccessListener(aVoid -> Log.v("Document Update", "Sucessfully updated Document"))
+                .addOnFailureListener(e -> Log.v("Document Update", "Something went wrong updating " + e.toString()));
     }
 }
